test : 로또 발급 역할 책임 변경으로 인한 테스트 코드 수정
