+++ src/test/java/lotto/domain/LottoCardTest.java
@@ -4,12 +4,12 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class LottoCardTest {
-
 	@DisplayName(value = "로또 카드의 1 ~ 45 번호가 생성되는지 테스트")
 	@Test
 	void lottoCardNumbers() {
@@ -21,5 +21,22 @@ class LottoCardTest {
 		}
 
 		assertThat(numbers).containsAll(expect);
+	}
+
+	@DisplayName(value = "로또 번호가 1부터 45까지 각각 다른 6개의 번호로 구성되는지 테스트")
+	@Test
+	void numbersOfLotto() {
+		LottoCard lottoCard = LottoCard.getInstance();
+		Lotto lotto = lottoCard.issue();
+
+		List<Integer> numbers = lotto.getNumbers()
+			.stream()
+			.distinct()
+			.collect(Collectors.toList());
+		int numberOfDistinctNumbers = numbers.size();
+		assertThat(numberOfDistinctNumbers).isEqualTo(6);
+
+		List<Integer> lottoCardNumbers = LottoCard.NUMBERS;
+		assertThat(lottoCardNumbers).containsAll(numbers);
 	}
 }
